lang: use macro for ToAccountMetas trait impls
