+++ lang/src/system_account.rs
@@ -28,17 +28,6 @@ impl<'info> SystemAccount<'info> {
 
 impl<'info> AccountsExit<'info> for SystemAccount<'info> {}
 
-impl<'info> ToAccountMetas for SystemAccount<'info> {
-    fn to_account_metas(&self, is_signer: Option<bool>) -> Vec<AccountMeta> {
-        let is_signer = is_signer.unwrap_or(self.info.is_signer);
-        let meta = match self.info.is_writable {
-            false => AccountMeta::new_readonly(*self.info.key, is_signer),
-            true => AccountMeta::new(*self.info.key, is_signer),
-        };
-        vec![meta]
-    }
-}
-
 impl<'info> Deref for SystemAccount<'info> {
     type Target = AccountInfo<'info>;
 
@@ -49,3 +38,4 @@ impl<'info> Deref for SystemAccount<'info> {
 
 impl_account_info_traits!(SystemAccount<'info>);
 impl_accounts_trait!(SystemAccount<'info>);
+impl_account_metas_trait!(SystemAccount<'info>);
