+++ lang/src/sysvar.rs
@@ -2,7 +2,6 @@ use crate::error::ErrorCode;
 use crate::{impl_account_info_traits, impl_accounts_trait};
 use crate::{Accounts, AccountsExit, Key, ToAccountInfo, ToAccountInfos, ToAccountMetas};
 use solana_program::account_info::AccountInfo;
-use solana_program::entrypoint::ProgramResult;
 use solana_program::instruction::AccountMeta;
 use solana_program::program_error::ProgramError;
 use solana_program::pubkey::Pubkey;
